Moved the main app under app/pillar.
